Add solution part2.
Found the solution on Reddit. I cannot get my own implementation going,so I tried the proposed solution, and that works. Might figure out later why it is not working

+++ Solutions/2023/Day24/Vector.cs
@@ -2,15 +2,15 @@
 {
     public class Vector
     {
-        public Vector(float x, float y, float z)
+        public Vector(double x, double y, double z)
         {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public float X { get; }
-        public float Y { get; }
-        public float Z { get; }
+        public double X { get; }
+        public double Y { get; }
+        public double Z { get; }
     }
 }
